CR FIX: Move abstractGenericAviationMessageParser to fi.fmi.avi.converter.iwxxm.generic
